src/requirements/base.txt
Django==2.0.5
django-material==1.2.5
pytz==2018.4
